using bootstrap grid; changed Vimeo.Players set-up
Moved away from our own implementation of a grid. Using Bootstrap
now, for responsiveness and effectiveness.

Also changed how the Vimeo.Players are set-up. Each video needs a
class of 'timeline-video' and a data-target with the id of the div
that represents the moving time and all is set-up automatically with
JS and jQuery.